Restructure access controller package for versioning
