+++ radix-engine-interface/src/blueprints/package/substates.rs
@@ -7,7 +7,7 @@ pub const RESOURCE_CODE_ID: u64 = 1u64;
 pub const IDENTITY_CODE_ID: u64 = 2u64;
 pub const CONSENSUS_MANAGER_CODE_ID: u64 = 3u64;
 pub const ACCOUNT_CODE_ID: u64 = 5u64;
-pub const ACCESS_CONTROLLER_CODE_ID: u64 = 6u64;
+pub const ACCESS_CONTROLLER_V1_0_CODE_ID: u64 = 6u64;
 pub const TRANSACTION_PROCESSOR_CODE_ID: u64 = 7u64;
 pub const METADATA_CODE_ID: u64 = 10u64;
 pub const ROYALTY_CODE_ID: u64 = 11u64;
