--- radix-engine/src/blueprints/access_controller/blueprint.rs
@@ -1,1250 +1,0 @@
-use super::events::*;
-use super::state_machine::*;
-use crate::errors::{ApplicationError, RuntimeError};
-use crate::internal_prelude::*;
-use crate::{event_schema, roles_template};
-use radix_blueprint_schema_init::{
-    BlueprintFunctionsSchemaInit, BlueprintSchemaInit, FunctionSchemaInit, ReceiverInfo, TypeRef,
-};
-use radix_common::time::Instant;
-use radix_engine_interface::api::field_api::LockFlags;
-use radix_engine_interface::api::object_api::ModuleId;
-use radix_engine_interface::blueprints::access_controller::*;
-use radix_engine_interface::blueprints::package::{
-    AuthConfig, BlueprintDefinitionInit, BlueprintType, FunctionAuth, MethodAuthTemplate,
-};
-use radix_engine_interface::blueprints::resource::*;
-use radix_engine_interface::object_modules::metadata::*;
-use radix_engine_interface::object_modules::ModuleConfig;
-use radix_engine_interface::*;
-use radix_engine_interface::{api::*, rule};
-use radix_native_sdk::modules::metadata::Metadata;
-use radix_native_sdk::modules::role_assignment::{
-    AttachedRoleAssignment, RoleAssignment, RoleAssignmentObject,
-};
-use radix_native_sdk::resource::NativeBucket;
-use radix_native_sdk::resource::NativeVault;
-use radix_native_sdk::runtime::Runtime;
-use sbor::rust::prelude::*;
-
-#[derive(Debug, PartialEq, Eq, ScryptoSbor)]
-pub struct AccessControllerSubstate {
-    /// A vault where the asset controlled by the access controller lives.
-    pub controlled_asset: Vault,
-
-    /// The amount of time (in minutes) that it takes for timed recovery to be done. Maximum is
-    /// 4,294,967,295 minutes which is 8171.5511700913 years. When this is [`None`], then timed
-    /// recovery can not be performed through this access controller.
-    pub timed_recovery_delay_in_minutes: Option<u32>,
-
-    /// The resource address of the recovery badge that will be used by the wallet and optionally
-    /// by other clients as well.
-    pub recovery_badge: ResourceAddress,
-
-    /// The states of the Access Controller.
-    pub state: (
-        // Controls whether the primary role is locked or unlocked
-        PrimaryRoleLockingState,
-        // Primary role recovery and withdraw states
-        PrimaryRoleRecoveryAttemptState,
-        PrimaryRoleBadgeWithdrawAttemptState,
-        // Recovery role recovery and withdraw states
-        RecoveryRoleRecoveryAttemptState,
-        RecoveryRoleBadgeWithdrawAttemptState,
-    ),
-}
-
-impl AccessControllerSubstate {
-    pub fn new(
-        controlled_asset: Vault,
-        timed_recovery_delay_in_minutes: Option<u32>,
-        recovery_badge: ResourceAddress,
-    ) -> Self {
-        Self {
-            controlled_asset,
-            timed_recovery_delay_in_minutes,
-            recovery_badge,
-            state: Default::default(),
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
-pub enum PrimaryRoleLockingState {
-    #[default]
-    Unlocked,
-    Locked,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
-pub enum PrimaryRoleRecoveryAttemptState {
-    #[default]
-    NoRecoveryAttempt,
-    RecoveryAttempt(RecoveryProposal),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
-pub enum PrimaryRoleBadgeWithdrawAttemptState {
-    #[default]
-    NoBadgeWithdrawAttempt,
-    BadgeWithdrawAttempt,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
-pub enum RecoveryRoleRecoveryAttemptState {
-    #[default]
-    NoRecoveryAttempt,
-    RecoveryAttempt(RecoveryRoleRecoveryState),
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
-pub enum RecoveryRoleRecoveryState {
-    UntimedRecovery(RecoveryProposal),
-    TimedRecovery {
-        proposal: RecoveryProposal,
-        timed_recovery_allowed_after: Instant,
-    },
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
-pub enum RecoveryRoleBadgeWithdrawAttemptState {
-    #[default]
-    NoBadgeWithdrawAttempt,
-    BadgeWithdrawAttempt,
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
-pub enum AccessControllerError {
-    /// Occurs when some action requires that the primary role is unlocked to happen.
-    OperationRequiresUnlockedPrimaryRole,
-
-    /// Occurs when adding time to an [`Instant`] results in an overflow
-    TimeOverflow,
-
-    /// Occurs when a proposer attempts to initiate another recovery when they already have a
-    /// recovery underway.
-    RecoveryAlreadyExistsForProposer { proposer: Proposer },
-
-    /// Occurs when no recovery can be found for a given proposer.
-    NoRecoveryExistsForProposer { proposer: Proposer },
-
-    /// Occurs when a proposer attempts to initiate another badge withdraw when they already have a
-    /// recovery underway.
-    BadgeWithdrawAttemptAlreadyExistsForProposer { proposer: Proposer },
-
-    /// Occurs when no recovery can be found for a given proposer.
-    NoBadgeWithdrawAttemptExistsForProposer { proposer: Proposer },
-
-    /// Occurs when there is no timed recoveries on the controller - typically because it isn't in
-    /// the state that allows for it.
-    NoTimedRecoveriesFound,
-
-    /// Occurs when trying to perform a timed confirm recovery on a recovery proposal that could
-    /// be time-confirmed but whose delay has not yet elapsed.
-    TimedRecoveryDelayHasNotElapsed,
-
-    /// Occurs when the expected recovery proposal doesn't match that which was found
-    RecoveryProposalMismatch {
-        expected: Box<RecoveryProposal>,
-        found: Box<RecoveryProposal>,
-    },
-}
-
-impl From<AccessControllerError> for RuntimeError {
-    fn from(value: AccessControllerError) -> Self {
-        RuntimeError::ApplicationError(ApplicationError::AccessControllerError(value))
-    }
-}
-
-declare_native_blueprint_state! {
-    blueprint_ident: AccessController,
-    blueprint_snake_case: access_controller,
-    features: {
-    },
-    fields: {
-        state:  {
-            ident: State,
-            field_type: {
-                kind: StaticSingleVersioned,
-            },
-            condition: Condition::Always,
-        }
-    },
-    collections: {
-    }
-}
-
-pub type AccessControllerStateV1 = AccessControllerSubstate;
-
-pub struct AccessControllerBlueprint;
-
-impl AccessControllerBlueprint {
-    pub fn definition() -> BlueprintDefinitionInit {
-        let mut aggregator = TypeAggregator::<ScryptoCustomTypeKind>::new();
-
-        let feature_set = AccessControllerFeatureSet::all_features();
-        let state = AccessControllerStateSchemaInit::create_schema_init(&mut aggregator);
-
-        let mut functions = index_map_new();
-        functions.insert(
-            ACCESS_CONTROLLER_CREATE_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: None,
-                input: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<AccessControllerCreateInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerCreateGlobalOutput>(),
-                ),
-                export: ACCESS_CONTROLLER_CREATE_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_CREATE_PROOF_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator.add_child_type_and_descendents::<AccessControllerCreateProofInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerCreateProofOutput>(),
-                ),
-                export: ACCESS_CONTROLLER_CREATE_PROOF_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_INITIATE_RECOVERY_AS_PRIMARY_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateRecoveryAsPrimaryInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateRecoveryAsPrimaryOutput>()),
-                export: ACCESS_CONTROLLER_INITIATE_RECOVERY_AS_PRIMARY_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_INITIATE_RECOVERY_AS_RECOVERY_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateRecoveryAsRecoveryInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateRecoveryAsRecoveryOutput>()),
-                export: ACCESS_CONTROLLER_INITIATE_RECOVERY_AS_RECOVERY_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_QUICK_CONFIRM_PRIMARY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmPrimaryRoleRecoveryProposalInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmPrimaryRoleRecoveryProposalOutput>()),
-                export: ACCESS_CONTROLLER_QUICK_CONFIRM_PRIMARY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_QUICK_CONFIRM_RECOVERY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmRecoveryRoleRecoveryProposalInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmRecoveryRoleRecoveryProposalOutput>()),
-                export: ACCESS_CONTROLLER_QUICK_CONFIRM_RECOVERY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_TIMED_CONFIRM_RECOVERY_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerTimedConfirmRecoveryInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerTimedConfirmRecoveryOutput>()),
-                export: ACCESS_CONTROLLER_TIMED_CONFIRM_RECOVERY_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_CANCEL_PRIMARY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelPrimaryRoleRecoveryProposalInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelPrimaryRoleRecoveryProposalOutput>()),
-                export: ACCESS_CONTROLLER_CANCEL_PRIMARY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_CANCEL_RECOVERY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelRecoveryRoleRecoveryProposalInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelRecoveryRoleRecoveryProposalOutput>()),
-                export: ACCESS_CONTROLLER_CANCEL_RECOVERY_ROLE_RECOVERY_PROPOSAL_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_LOCK_PRIMARY_ROLE_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerLockPrimaryRoleInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerLockPrimaryRoleOutput>(),
-                ),
-                export: ACCESS_CONTROLLER_LOCK_PRIMARY_ROLE_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_UNLOCK_PRIMARY_ROLE_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerUnlockPrimaryRoleInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerUnlockPrimaryRoleOutput>(
-                        ),
-                ),
-                export: ACCESS_CONTROLLER_UNLOCK_PRIMARY_ROLE_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_STOP_TIMED_RECOVERY_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerStopTimedRecoveryInput>(),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerStopTimedRecoveryOutput>(
-                        ),
-                ),
-                export: ACCESS_CONTROLLER_STOP_TIMED_RECOVERY_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_INITIATE_BADGE_WITHDRAW_ATTEMPT_AS_PRIMARY_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateBadgeWithdrawAttemptAsPrimaryInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateBadgeWithdrawAttemptAsPrimaryOutput>()),
-                export: ACCESS_CONTROLLER_INITIATE_BADGE_WITHDRAW_ATTEMPT_AS_PRIMARY_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_INITIATE_BADGE_WITHDRAW_ATTEMPT_AS_RECOVERY_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateBadgeWithdrawAttemptAsRecoveryInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerInitiateBadgeWithdrawAttemptAsRecoveryOutput>()),
-                export: ACCESS_CONTROLLER_INITIATE_BADGE_WITHDRAW_ATTEMPT_AS_RECOVERY_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_QUICK_CONFIRM_PRIMARY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmPrimaryRoleBadgeWithdrawAttemptInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmPrimaryRoleBadgeWithdrawAttemptOutput>()),
-                export: ACCESS_CONTROLLER_QUICK_CONFIRM_PRIMARY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_QUICK_CONFIRM_RECOVERY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmRecoveryRoleBadgeWithdrawAttemptInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerQuickConfirmRecoveryRoleBadgeWithdrawAttemptOutput>()),
-                export: ACCESS_CONTROLLER_QUICK_CONFIRM_RECOVERY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_CANCEL_PRIMARY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelPrimaryRoleBadgeWithdrawAttemptInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelPrimaryRoleBadgeWithdrawAttemptOutput>()),
-                export: ACCESS_CONTROLLER_CANCEL_PRIMARY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_CANCEL_RECOVERY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelRecoveryRoleBadgeWithdrawAttemptInput>()),
-                output: TypeRef::Static(aggregator
-                    .add_child_type_and_descendents::<AccessControllerCancelRecoveryRoleBadgeWithdrawAttemptOutput>()),
-                export: ACCESS_CONTROLLER_CANCEL_RECOVERY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT.to_string(),
-            },
-        );
-        functions.insert(
-            ACCESS_CONTROLLER_MINT_RECOVERY_BADGES_IDENT.to_string(),
-            FunctionSchemaInit {
-                receiver: Some(ReceiverInfo::normal_ref_mut()),
-                input: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerMintRecoveryBadgesInput>(
-                        ),
-                ),
-                output: TypeRef::Static(
-                    aggregator
-                        .add_child_type_and_descendents::<AccessControllerMintRecoveryBadgesOutput>(
-                        ),
-                ),
-                export: ACCESS_CONTROLLER_MINT_RECOVERY_BADGES_IDENT.to_string(),
-            },
-        );
-
-        let events = event_schema! {
-            aggregator,
-            [
-                InitiateRecoveryEvent,
-                RuleSetUpdateEvent,
-                CancelRecoveryProposalEvent,
-                LockPrimaryRoleEvent,
-                UnlockPrimaryRoleEvent,
-                StopTimedRecoveryEvent,
-                InitiateBadgeWithdrawAttemptEvent,
-                BadgeWithdrawEvent,
-                CancelBadgeWithdrawAttemptEvent
-            ]
-        };
-
-        let schema = generate_full_schema(aggregator);
-        BlueprintDefinitionInit {
-            blueprint_type: BlueprintType::default(),
-            is_transient: false,
-            feature_set,
-            dependencies: indexset!(PACKAGE_OF_DIRECT_CALLER_VIRTUAL_BADGE.into(),),
-
-            schema: BlueprintSchemaInit {
-                generics: vec![],
-                schema,
-                state,
-                events,
-                types: BlueprintTypeSchemaInit::default(),
-                functions: BlueprintFunctionsSchemaInit { functions },
-                hooks: BlueprintHooksInit::default(),
-            },
-
-            royalty_config: PackageRoyaltyConfig::default(),
-            auth_config: AuthConfig {
-                function_auth: FunctionAuth::AllowAll,
-                method_auth: MethodAuthTemplate::StaticRoleDefinition(roles_template!(
-                    roles {
-                        "primary" => updaters: [SELF_ROLE];
-                        "recovery" => updaters: [SELF_ROLE];
-                        "confirmation" => updaters: [SELF_ROLE];
-                    },
-                    methods {
-                        ACCESS_CONTROLLER_TIMED_CONFIRM_RECOVERY_IDENT => MethodAccessibility::Public;
-
-                        ACCESS_CONTROLLER_CREATE_PROOF_IDENT => ["primary"];
-
-                        ACCESS_CONTROLLER_INITIATE_RECOVERY_AS_PRIMARY_IDENT => ["primary"];
-                        ACCESS_CONTROLLER_CANCEL_PRIMARY_ROLE_RECOVERY_PROPOSAL_IDENT => ["primary"];
-                        ACCESS_CONTROLLER_INITIATE_BADGE_WITHDRAW_ATTEMPT_AS_PRIMARY_IDENT => ["primary"];
-                        ACCESS_CONTROLLER_CANCEL_PRIMARY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT =>  ["primary"];
-
-                        ACCESS_CONTROLLER_INITIATE_RECOVERY_AS_RECOVERY_IDENT => ["recovery"];
-                        ACCESS_CONTROLLER_CANCEL_RECOVERY_ROLE_RECOVERY_PROPOSAL_IDENT => ["recovery"];
-                        ACCESS_CONTROLLER_INITIATE_BADGE_WITHDRAW_ATTEMPT_AS_RECOVERY_IDENT => ["recovery"];
-                        ACCESS_CONTROLLER_CANCEL_RECOVERY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT => ["recovery"];
-
-                        ACCESS_CONTROLLER_LOCK_PRIMARY_ROLE_IDENT => ["recovery"];
-                        ACCESS_CONTROLLER_UNLOCK_PRIMARY_ROLE_IDENT => ["recovery"];
-
-                        ACCESS_CONTROLLER_QUICK_CONFIRM_PRIMARY_ROLE_RECOVERY_PROPOSAL_IDENT => ["recovery", "confirmation"];
-                        ACCESS_CONTROLLER_QUICK_CONFIRM_PRIMARY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT => ["recovery", "confirmation"];
-
-                        ACCESS_CONTROLLER_QUICK_CONFIRM_RECOVERY_ROLE_RECOVERY_PROPOSAL_IDENT => ["primary", "confirmation"];
-                        ACCESS_CONTROLLER_QUICK_CONFIRM_RECOVERY_ROLE_BADGE_WITHDRAW_ATTEMPT_IDENT => ["primary", "confirmation"];
-
-                        ACCESS_CONTROLLER_MINT_RECOVERY_BADGES_IDENT => ["primary", "recovery"];
-
-                        ACCESS_CONTROLLER_STOP_TIMED_RECOVERY_IDENT => ["primary", "confirmation", "recovery"];
-                    }
-                )),
-            },
-        }
-    }
-
-    pub fn create<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let input: AccessControllerCreateInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        // Allocating the address of the access controller - this will be needed for the metadata
-        // and access rules of the recovery badge
-        let (address_reservation, address) = {
-            if let Some(address_reservation) = input.address_reservation {
-                let address = api.get_reservation_address(address_reservation.0.as_node_id())?;
-                (address_reservation, address)
-            } else {
-                api.allocate_global_address(BlueprintId {
-                    package_address: ACCESS_CONTROLLER_PACKAGE,
-                    blueprint_name: ACCESS_CONTROLLER_BLUEPRINT.to_string(),
-                })?
-            }
-        };
-
-        // Creating a new vault and putting in it the controlled asset
-        let vault = {
-            let mut vault = input
-                .controlled_asset
-                .resource_address(api)
-                .and_then(|resource_address| Vault::create(resource_address, api))?;
-            vault.put(input.controlled_asset, api)?;
-
-            vault
-        };
-
-        // Creating a new recovery badge resource
-        let recovery_badge_resource = {
-            let global_component_caller_badge =
-                NonFungibleGlobalId::global_caller_badge(GlobalCaller::GlobalObject(address));
-
-            let resource_address = {
-                let non_fungible_schema =
-                    NonFungibleDataSchema::new_local_without_self_package_replacement::<()>();
-
-                let result = api.call_function(
-                    RESOURCE_PACKAGE,
-                    NON_FUNGIBLE_RESOURCE_MANAGER_BLUEPRINT,
-                    NON_FUNGIBLE_RESOURCE_MANAGER_CREATE_IDENT,
-                    scrypto_encode(&NonFungibleResourceManagerCreateInput {
-                        owner_role: OwnerRole::Fixed(rule!(require(global_component_caller_badge.clone()))),
-                        id_type: NonFungibleIdType::Integer,
-                        track_total_supply: true,
-                        non_fungible_schema,
-                        resource_roles: NonFungibleResourceRoles {
-                            mint_roles: mint_roles! {
-                                minter => rule!(require(global_component_caller_badge.clone()));
-                                minter_updater => rule!(deny_all);
-                            },
-                            burn_roles: burn_roles! {
-                                burner => rule!(allow_all);
-                                burner_updater => rule!(allow_all);
-                            },
-                            withdraw_roles: withdraw_roles! {
-                                withdrawer => rule!(deny_all);
-                                withdrawer_updater => rule!(deny_all);
-                            },
-                            ..Default::default()
-                        },
-                        metadata: metadata! {
-                            roles {
-                                metadata_setter => AccessRule::DenyAll;
-                                metadata_setter_updater => AccessRule::DenyAll;
-                                metadata_locker => AccessRule::DenyAll;
-                                metadata_locker_updater => AccessRule::DenyAll;
-                            },
-                            init {
-                                "name" => "Recovery Badge".to_owned(), locked;
-                                "icon_url" => UncheckedUrl::of("https://assets.radixdlt.com/icons/icon-recovery_badge.png".to_owned()), locked;
-                                "access_controller" => address, locked;
-                            }
-                        },
-                        address_reservation: None,
-                    })
-                        .unwrap(),
-                )?;
-                scrypto_decode::<ResourceAddress>(result.as_slice()).unwrap()
-            };
-
-            resource_address
-        };
-
-        let substate = AccessControllerSubstate::new(
-            vault,
-            input.timed_recovery_delay_in_minutes,
-            recovery_badge_resource,
-        );
-        let object_id = api.new_simple_object(
-            ACCESS_CONTROLLER_BLUEPRINT,
-            indexmap! {
-                AccessControllerField::State.field_index() => FieldValue::new(&AccessControllerStateFieldPayload::from_content_source(substate)),
-            },
-        )?;
-
-        let roles = init_roles_from_rule_set(input.rule_set);
-        let roles = indexmap!(ModuleId::Main => roles);
-        let role_assignment = RoleAssignment::create(OwnerRole::None, roles, api)?.0;
-
-        let metadata = Metadata::create_with_data(
-            metadata_init! {
-                "recovery_badge" => GlobalAddress::from(recovery_badge_resource), locked;
-            },
-            api,
-        )?;
-
-        // Creating a global component address for the access controller RENode
-        api.globalize(
-            object_id,
-            indexmap!(
-                AttachedModuleId::RoleAssignment => role_assignment.0,
-                AttachedModuleId::Metadata => metadata.0,
-            ),
-            Some(address_reservation),
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&address))
-    }
-
-    pub fn create_proof<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let _input: AccessControllerCreateProofInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        let proof = transition(api, AccessControllerCreateProofStateMachineInput)?;
-
-        Ok(IndexedScryptoValue::from_typed(&proof))
-    }
-
-    pub fn initiate_recovery_as_primary<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let input: AccessControllerInitiateRecoveryAsPrimaryInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-        let proposal = RecoveryProposal {
-            rule_set: input.rule_set,
-            timed_recovery_delay_in_minutes: input.timed_recovery_delay_in_minutes,
-        };
-
-        transition_mut(
-            api,
-            AccessControllerInitiateRecoveryAsPrimaryStateMachineInput {
-                proposal: proposal.clone(),
-            },
-        )?;
-
-        Runtime::emit_event(
-            api,
-            InitiateRecoveryEvent {
-                proposal,
-                proposer: Proposer::Primary,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn initiate_recovery_as_recovery<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let input: AccessControllerInitiateRecoveryAsRecoveryInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-        let proposal = RecoveryProposal {
-            rule_set: input.rule_set,
-            timed_recovery_delay_in_minutes: input.timed_recovery_delay_in_minutes,
-        };
-
-        transition_mut(
-            api,
-            AccessControllerInitiateRecoveryAsRecoveryStateMachineInput {
-                proposal: proposal.clone(),
-            },
-        )?;
-
-        Runtime::emit_event(
-            api,
-            InitiateRecoveryEvent {
-                proposal,
-                proposer: Proposer::Recovery,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn initiate_badge_withdraw_attempt_as_primary<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        input
-            .as_typed::<AccessControllerInitiateBadgeWithdrawAttemptAsPrimaryInput>()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(
-            api,
-            AccessControllerInitiateBadgeWithdrawAttemptAsPrimaryStateMachineInput,
-        )?;
-
-        Runtime::emit_event(
-            api,
-            InitiateBadgeWithdrawAttemptEvent {
-                proposer: Proposer::Primary,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn initiate_badge_withdraw_attempt_as_recovery<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        input
-            .as_typed::<AccessControllerInitiateBadgeWithdrawAttemptAsRecoveryInput>()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(
-            api,
-            AccessControllerInitiateBadgeWithdrawAttemptAsRecoveryStateMachineInput,
-        )?;
-
-        Runtime::emit_event(
-            api,
-            InitiateBadgeWithdrawAttemptEvent {
-                proposer: Proposer::Recovery,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn quick_confirm_primary_role_recovery_proposal<Y>(
-        receiver: &NodeId,
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let input: AccessControllerQuickConfirmPrimaryRoleRecoveryProposalInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-        let proposal = RecoveryProposal {
-            rule_set: input.rule_set,
-            timed_recovery_delay_in_minutes: input.timed_recovery_delay_in_minutes,
-        };
-
-        let recovery_proposal = transition_mut(
-            api,
-            AccessControllerQuickConfirmPrimaryRoleRecoveryProposalStateMachineInput {
-                proposal_to_confirm: proposal.clone(),
-            },
-        )?;
-
-        update_role_assignment(api, receiver, recovery_proposal.rule_set)?;
-
-        Runtime::emit_event(
-            api,
-            RuleSetUpdateEvent {
-                proposal,
-                proposer: Proposer::Primary,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn quick_confirm_recovery_role_recovery_proposal<Y>(
-        receiver: &NodeId,
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let input: AccessControllerQuickConfirmRecoveryRoleRecoveryProposalInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-        let proposal = RecoveryProposal {
-            rule_set: input.rule_set,
-            timed_recovery_delay_in_minutes: input.timed_recovery_delay_in_minutes,
-        };
-
-        let recovery_proposal = transition_mut(
-            api,
-            AccessControllerQuickConfirmRecoveryRoleRecoveryProposalStateMachineInput {
-                proposal_to_confirm: proposal.clone(),
-            },
-        )?;
-
-        update_role_assignment(api, receiver, recovery_proposal.rule_set)?;
-
-        Runtime::emit_event(
-            api,
-            RuleSetUpdateEvent {
-                proposal,
-                proposer: Proposer::Recovery,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn quick_confirm_primary_role_badge_withdraw_attempt<Y>(
-        receiver: &NodeId,
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        input
-            .as_typed::<AccessControllerQuickConfirmPrimaryRoleBadgeWithdrawAttemptInput>()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        let bucket = transition_mut(
-            api,
-            AccessControllerQuickConfirmPrimaryRoleBadgeWithdrawAttemptStateMachineInput,
-        )?;
-
-        update_role_assignment(api, receiver, locked_role_assignment())?;
-
-        Runtime::emit_event(
-            api,
-            BadgeWithdrawEvent {
-                proposer: Proposer::Primary,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&bucket))
-    }
-
-    pub fn quick_confirm_recovery_role_badge_withdraw_attempt<Y>(
-        receiver: &NodeId,
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        input
-            .as_typed::<AccessControllerQuickConfirmRecoveryRoleBadgeWithdrawAttemptInput>()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        let bucket = transition_mut(
-            api,
-            AccessControllerQuickConfirmRecoveryRoleBadgeWithdrawAttemptStateMachineInput,
-        )?;
-
-        update_role_assignment(api, receiver, locked_role_assignment())?;
-
-        Runtime::emit_event(
-            api,
-            BadgeWithdrawEvent {
-                proposer: Proposer::Recovery,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&bucket))
-    }
-
-    pub fn timed_confirm_recovery<Y>(
-        receiver: &NodeId,
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let input: AccessControllerTimedConfirmRecoveryInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-        let proposal = RecoveryProposal {
-            rule_set: input.rule_set,
-            timed_recovery_delay_in_minutes: input.timed_recovery_delay_in_minutes,
-        };
-
-        let recovery_proposal = transition_mut(
-            api,
-            AccessControllerTimedConfirmRecoveryStateMachineInput {
-                proposal_to_confirm: proposal.clone(),
-            },
-        )?;
-
-        // Update the access rules
-        update_role_assignment(api, receiver, recovery_proposal.rule_set)?;
-
-        Runtime::emit_event(
-            api,
-            RuleSetUpdateEvent {
-                proposal,
-                proposer: Proposer::Recovery,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn cancel_primary_role_recovery_proposal<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let _input: AccessControllerCancelPrimaryRoleRecoveryProposalInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(
-            api,
-            AccessControllerCancelPrimaryRoleRecoveryProposalStateMachineInput,
-        )?;
-
-        Runtime::emit_event(
-            api,
-            CancelRecoveryProposalEvent {
-                proposer: Proposer::Primary,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn cancel_recovery_role_recovery_proposal<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let _input: AccessControllerCancelRecoveryRoleRecoveryProposalInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(
-            api,
-            AccessControllerCancelRecoveryRoleRecoveryProposalStateMachineInput,
-        )?;
-
-        Runtime::emit_event(
-            api,
-            CancelRecoveryProposalEvent {
-                proposer: Proposer::Recovery,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn cancel_primary_role_badge_withdraw_attempt<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        input
-            .as_typed::<AccessControllerCancelPrimaryRoleBadgeWithdrawAttemptInput>()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(
-            api,
-            AccessControllerCancelPrimaryRoleBadgeWithdrawAttemptStateMachineInput,
-        )?;
-
-        Runtime::emit_event(
-            api,
-            CancelBadgeWithdrawAttemptEvent {
-                proposer: Proposer::Primary,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn cancel_recovery_role_badge_withdraw_attempt<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        input
-            .as_typed::<AccessControllerCancelRecoveryRoleBadgeWithdrawAttemptInput>()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(
-            api,
-            AccessControllerCancelRecoveryRoleBadgeWithdrawAttemptStateMachineInput,
-        )?;
-
-        Runtime::emit_event(
-            api,
-            CancelBadgeWithdrawAttemptEvent {
-                proposer: Proposer::Recovery,
-            },
-        )?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn lock_primary_role<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let _input: AccessControllerLockPrimaryRoleInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(api, AccessControllerLockPrimaryRoleStateMachineInput)?;
-        Runtime::emit_event(api, LockPrimaryRoleEvent {})?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn unlock_primary_role<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let _input: AccessControllerUnlockPrimaryRoleInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(api, AccessControllerUnlockPrimaryRoleStateMachineInput)?;
-        Runtime::emit_event(api, UnlockPrimaryRoleEvent {})?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn stop_timed_recovery<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let input: AccessControllerStopTimedRecoveryInput = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        transition_mut(
-            api,
-            AccessControllerStopTimedRecoveryStateMachineInput {
-                proposal: RecoveryProposal {
-                    rule_set: input.rule_set,
-                    timed_recovery_delay_in_minutes: input.timed_recovery_delay_in_minutes,
-                },
-            },
-        )?;
-        Runtime::emit_event(api, StopTimedRecoveryEvent)?;
-
-        Ok(IndexedScryptoValue::from_typed(&()))
-    }
-
-    pub fn mint_recovery_badges<Y>(
-        input: &IndexedScryptoValue,
-        api: &mut Y,
-    ) -> Result<IndexedScryptoValue, RuntimeError>
-    where
-        Y: ClientApi<RuntimeError>,
-    {
-        let AccessControllerMintRecoveryBadgesInput {
-            non_fungible_local_ids,
-        } = input
-            .as_typed()
-            .map_err(|e| RuntimeError::ApplicationError(ApplicationError::InputDecodeError(e)))?;
-
-        let resource_address = {
-            let handle = api.actor_open_field(
-                ACTOR_STATE_SELF,
-                AccessControllerField::State.field_index(),
-                LockFlags::read_only(),
-            )?;
-
-            let access_controller = {
-                let access_controller: AccessControllerStateFieldPayload =
-                    api.field_read_typed(handle)?;
-                access_controller.into_latest()
-            };
-            access_controller.recovery_badge
-        };
-
-        let non_fungibles: IndexMap<NonFungibleLocalId, (ScryptoValue,)> = non_fungible_local_ids
-            .into_iter()
-            .map(|local_id| {
-                (
-                    local_id,
-                    (scrypto_decode(&scrypto_encode(&()).unwrap()).unwrap(),),
-                )
-            })
-            .collect();
-
-        let rtn = api.call_method(
-            resource_address.as_node_id(),
-            NON_FUNGIBLE_RESOURCE_MANAGER_MINT_IDENT,
-            scrypto_encode(&NonFungibleResourceManagerMintInput {
-                entries: non_fungibles,
-            })
-            .unwrap(),
-        )?;
-
-        Ok(IndexedScryptoValue::from_slice(&rtn).unwrap())
-    }
-}
-
-//=========
-// Helpers
-//=========
-
-fn locked_role_assignment() -> RuleSet {
-    RuleSet {
-        primary_role: AccessRule::DenyAll,
-        recovery_role: AccessRule::DenyAll,
-        confirmation_role: AccessRule::DenyAll,
-    }
-}
-
-fn init_roles_from_rule_set(rule_set: RuleSet) -> RoleAssignmentInit {
-    roles2! {
-        "primary" => rule_set.primary_role, updatable;
-        "recovery" => rule_set.recovery_role, updatable;
-        "confirmation" => rule_set.confirmation_role, updatable;
-    }
-}
-
-fn transition<Y, I>(
-    api: &mut Y,
-    input: I,
-) -> Result<<AccessControllerSubstate as Transition<I>>::Output, RuntimeError>
-where
-    Y: ClientApi<RuntimeError>,
-    AccessControllerSubstate: Transition<I>,
-{
-    let handle = api.actor_open_field(
-        ACTOR_STATE_SELF,
-        AccessControllerField::State.field_index(),
-        LockFlags::read_only(),
-    )?;
-
-    let access_controller = {
-        let access_controller: AccessControllerStateFieldPayload = api.field_read_typed(handle)?;
-        access_controller.into_latest()
-    };
-
-    let rtn = access_controller.transition(api, input)?;
-
-    api.field_close(handle)?;
-
-    Ok(rtn)
-}
-
-fn transition_mut<Y, I>(
-    api: &mut Y,
-    input: I,
-) -> Result<<AccessControllerSubstate as TransitionMut<I>>::Output, RuntimeError>
-where
-    Y: ClientApi<RuntimeError>,
-    AccessControllerSubstate: TransitionMut<I>,
-{
-    let handle = api.actor_open_field(
-        ACTOR_STATE_SELF,
-        AccessControllerField::State.field_index(),
-        LockFlags::MUTABLE,
-    )?;
-
-    let mut access_controller = {
-        let access_controller: AccessControllerStateFieldPayload = api.field_read_typed(handle)?;
-        access_controller.into_latest()
-    };
-
-    let rtn = access_controller.transition_mut(api, input)?;
-
-    {
-        api.field_write_typed(
-            handle,
-            &AccessControllerStateFieldPayload::from_content_source(access_controller),
-        )?;
-    }
-
-    api.field_close(handle)?;
-
-    Ok(rtn)
-}
-
-fn update_role_assignment<Y>(
-    api: &mut Y,
-    receiver: &NodeId,
-    rule_set: RuleSet,
-) -> Result<(), RuntimeError>
-where
-    Y: ClientApi<RuntimeError>,
-{
-    let attached = AttachedRoleAssignment(receiver.clone());
-    attached.set_role(
-        ModuleId::Main,
-        RoleKey::new("primary"),
-        rule_set.primary_role.clone(),
-        api,
-    )?;
-    attached.set_role(
-        ModuleId::Main,
-        RoleKey::new("recovery"),
-        rule_set.recovery_role.clone(),
-        api,
-    )?;
-    attached.set_role(
-        ModuleId::Main,
-        RoleKey::new("confirmation"),
-        rule_set.confirmation_role.clone(),
-        api,
-    )?;
-
-    Ok(())
-}
