+++ radix-engine/src/blueprints/access_controller/mod.rs
@@ -1,8 +1,1 @@
-mod blueprint;
-mod events;
-mod package;
-mod state_machine;
-
-pub use blueprint::*;
-pub use events::*;
-pub use package::*;
+pub mod v1;
