+++ radix-engine/src/blueprints/access_controller/v1/state_common.rs
@@ -1,0 +1,46 @@
+use radix_common::prelude::*;
+use radix_engine_interface::blueprints::access_controller::*;
+
+#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
+pub enum PrimaryRoleLockingState {
+    #[default]
+    Unlocked,
+    Locked,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
+pub enum PrimaryRoleRecoveryAttemptState {
+    #[default]
+    NoRecoveryAttempt,
+    RecoveryAttempt(RecoveryProposal),
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
+pub enum PrimaryRoleBadgeWithdrawAttemptState {
+    #[default]
+    NoBadgeWithdrawAttempt,
+    BadgeWithdrawAttempt,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
+pub enum RecoveryRoleRecoveryAttemptState {
+    #[default]
+    NoRecoveryAttempt,
+    RecoveryAttempt(RecoveryRoleRecoveryState),
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor)]
+pub enum RecoveryRoleRecoveryState {
+    UntimedRecovery(RecoveryProposal),
+    TimedRecovery {
+        proposal: RecoveryProposal,
+        timed_recovery_allowed_after: Instant,
+    },
+}
+
+#[derive(Debug, Clone, PartialEq, Eq, ScryptoSbor, Default)]
+pub enum RecoveryRoleBadgeWithdrawAttemptState {
+    #[default]
+    NoBadgeWithdrawAttempt,
+    BadgeWithdrawAttempt,
+}
