+++ radix-engine/src/blueprints/native_schema.rs
@@ -1,7 +1,7 @@
 use lazy_static::lazy_static;
 use radix_engine_interface::blueprints::package::*;
 
-use crate::blueprints::access_controller::*;
+use crate::blueprints::access_controller::v1::*;
 use crate::blueprints::account::*;
 use crate::blueprints::consensus_manager::*;
 use crate::blueprints::identity::*;
@@ -23,7 +23,7 @@ lazy_static! {
     pub static ref IDENTITY_PACKAGE_DEFINITION: PackageDefinition =
         IdentityNativePackage::definition();
     pub static ref ACCESS_CONTROLLER_PACKAGE_DEFINITION: PackageDefinition =
-        AccessControllerNativePackage::definition();
+        AccessControllerV1NativePackage::definition();
     pub static ref POOL_PACKAGE_DEFINITION_V1_0: PackageDefinition =
         PoolNativePackage::definition(PoolV1MinorVersion::Zero);
     pub static ref POOL_PACKAGE_DEFINITION_V1_1: PackageDefinition =
