+++ radix-substate-store-queries/src/typed_substate_layout.rs
@@ -6,7 +6,7 @@ use radix_engine_interface::prelude::*;
 
 // Import and re-export these types so they are available easily with a single import
 use radix_common::prelude::*;
-pub use radix_engine::blueprints::access_controller::*;
+pub use radix_engine::blueprints::access_controller::v1::v1_0 as access_controller_v1_0;
 pub use radix_engine::blueprints::account::{AccountBlueprint, AccountError, AccountNativePackage};
 use radix_engine::blueprints::account::{AccountTypedSubstateKey, AccountTypedSubstateValue};
 pub use radix_engine::blueprints::consensus_manager::*;
@@ -153,7 +153,7 @@ pub enum TypedMainModuleSubstateKey {
     NonFungibleVault(NonFungibleVaultTypedSubstateKey),
     ConsensusManager(ConsensusManagerTypedSubstateKey),
     ValidatorField(ValidatorTypedSubstateKey),
-    AccessController(AccessControllerTypedSubstateKey),
+    AccessControllerV1(access_controller_v1_0::AccessControllerTypedSubstateKey),
     Account(AccountTypedSubstateKey),
     OneResourcePool(OneResourcePoolTypedSubstateKey),
     TwoResourcePool(TwoResourcePoolTypedSubstateKey),
@@ -301,8 +301,8 @@ fn to_typed_object_substate_key_internal(
         EntityType::GlobalValidator => TypedMainModuleSubstateKey::ValidatorField(
             ValidatorTypedSubstateKey::for_key_at_partition_offset(partition_offset, substate_key)?,
         ),
-        EntityType::GlobalAccessController => TypedMainModuleSubstateKey::AccessController(
-            AccessControllerTypedSubstateKey::for_key_in_partition(
+        EntityType::GlobalAccessController => TypedMainModuleSubstateKey::AccessControllerV1(
+            access_controller_v1_0::AccessControllerTypedSubstateKey::for_key_in_partition(
                 &AccessControllerPartitionOffset::try_from(partition_offset)?,
                 substate_key,
             )?,
@@ -428,7 +428,7 @@ pub enum TypedMainModuleSubstateValue {
     NonFungibleVault(NonFungibleVaultTypedSubstateValue),
     ConsensusManager(ConsensusManagerTypedSubstateValue),
     Validator(ValidatorTypedSubstateValue),
-    AccessController(AccessControllerTypedSubstateValue),
+    AccessControllerV1(access_controller_v1_0::AccessControllerTypedSubstateValue),
     Account(AccountTypedSubstateValue),
     OneResourcePool(OneResourcePoolTypedSubstateValue),
     TwoResourcePool(TwoResourcePoolTypedSubstateValue),
@@ -560,9 +560,11 @@ fn to_typed_object_substate_value(
         TypedMainModuleSubstateKey::Account(key) => TypedMainModuleSubstateValue::Account(
             AccountTypedSubstateValue::from_key_and_data(key, data)?,
         ),
-        TypedMainModuleSubstateKey::AccessController(key) => {
-            TypedMainModuleSubstateValue::AccessController(
-                AccessControllerTypedSubstateValue::from_key_and_data(key, data)?,
+        TypedMainModuleSubstateKey::AccessControllerV1(key) => {
+            TypedMainModuleSubstateValue::AccessControllerV1(
+                access_controller_v1_0::AccessControllerTypedSubstateValue::from_key_and_data(
+                    key, data,
+                )?,
             )
         }
         TypedMainModuleSubstateKey::GenericScryptoComponentField(offset) => {
